backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
yfinance==0.2.12
pandas==1.5.3
numpy==1.24.3
pymongo==4.3.3
python-dotenv==1.0.0
pytest==7.3.1
matplotlib==3.7.1
